pyproject.toml
[tool.poetry]
name = "scripts"
version = "0.1.0"
description = ""
authors = ["Shuanghui Bao <baoshrainbow@126.com>"]
readme = "README.md"
#package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
openai = "^1.54.4"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

packages = [
    { include = "scripts" },
]

[tool.poetry.scripts]
start = "scripts.main:app_run"
test = "scripts.main:app_test"
